refactor: split monolithic MCP implementation into focused files
Extract the 639-line mcp.go into 7 files by responsibility:
protocol types, tool definitions, each tool handler, and shared utilities.
Replace chained if-else dispatch with switch statement and deduplicate
exec.Command boilerplate into runCommand/runCommandRaw helpers.
